requirements.txt
vanna>=0.0.30
psycopg2-binary>=2.9.9
openai>=1.12.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0 